Remove jax-md and rewrite periodic.py (#1)
This commit removes the dependency on jax-md.

To do this, I had to rewrite periodic.py, which
now replicates the (little) functionatlity we
previously imported from jax-md. The core here
is the implementation of how to compute MIC
displacements between two positions. In this,
I take a different route than jax-md:

- cell is an explicit argument
- different sign convention
- no custom jvp for the transform into fractional coords

As a result of the last change, the gradients
of the new displacements with respect to the cell
are now correct. Therefore, we can compute the stress
by directly differentiating with respect to positions
and cell, which was not previously possible.

This bumps the version to 0.2.0.