Jittable Unfolding
This makes the unfolding procedure jittable.

We adopt the pattern set by neighborlist, adding
some dummy padding positions to keep shapes constants.
With this, we can move the whole procedure of computing
how to unfolding into the JIT, and avoid the extremely
high costs incorred by overflows.

We pay a price in complexity for this: now the unfolded system
contains some fake positions that have to be removed somehow.
The solution I've chosen is to do this in the neighborlist, which
becomes more complicated, as we now have to pass the mask around.